backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
alembic==1.13.1
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
confluent-kafka==2.3.0
loguru==0.7.2
structlog==24.1.0
prometheus-client==0.20.0
python-dotenv==1.0.1
pywin32
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.27.0
